Move card tags to mod-scope class
