+++ src/main/java/stsjorbsmod/cards/wanderer/ArcaneForm.java
@@ -10,8 +10,8 @@ import stsjorbsmod.characters.Wanderer;
 import stsjorbsmod.memories.*;
 
 import static stsjorbsmod.JorbsMod.makeCardPath;
-import static stsjorbsmod.characters.Wanderer.Enums.PERSISTENT_POSITIVE_EFFECT;
-import static stsjorbsmod.characters.Wanderer.Enums.REMEMBER_MEMORY;
+import static stsjorbsmod.JorbsMod.JorbsCardTags.PERSISTENT_POSITIVE_EFFECT;
+import static stsjorbsmod.JorbsMod.JorbsCardTags.REMEMBER_MEMORY;
 
 public class ArcaneForm extends CustomJorbsModCard {
     public static final String ID = JorbsMod.makeID(ArcaneForm.class.getSimpleName());
