+++ src/main/java/stsjorbsmod/cards/wanderer/ColorSpray.java
@@ -16,7 +16,7 @@ import stsjorbsmod.powers.BanishedPower;
 import stsjorbsmod.powers.BurningPower;
 
 import static stsjorbsmod.JorbsMod.makeCardPath;
-import static stsjorbsmod.characters.Wanderer.Enums.REMEMBER_MEMORY;
+import static stsjorbsmod.JorbsMod.JorbsCardTags.REMEMBER_MEMORY;
 
 public class ColorSpray extends CustomJorbsModCard {
     public static final String ID = JorbsMod.makeID(ColorSpray.class.getSimpleName());
