+++ src/main/java/stsjorbsmod/cards/wanderer/EyeOfTheStorm.java
@@ -10,7 +10,7 @@ import stsjorbsmod.actions.RememberRandomNewMemoryAction;
 import stsjorbsmod.characters.Wanderer;
 
 import static stsjorbsmod.JorbsMod.makeCardPath;
-import static stsjorbsmod.characters.Wanderer.Enums.REMEMBER_MEMORY;
+import static stsjorbsmod.JorbsMod.JorbsCardTags.REMEMBER_MEMORY;
 
 public class EyeOfTheStorm extends CustomJorbsModCard {
     public static final String ID = JorbsMod.makeID(EyeOfTheStorm.class.getSimpleName());
