+++ src/main/java/stsjorbsmod/cards/wanderer/Hibernate.java
@@ -10,7 +10,7 @@ import stsjorbsmod.characters.Wanderer;
 import stsjorbsmod.memories.SlothMemory;
 
 import static stsjorbsmod.JorbsMod.makeCardPath;
-import static stsjorbsmod.characters.Wanderer.Enums.REMEMBER_MEMORY;
+import static stsjorbsmod.JorbsMod.JorbsCardTags.REMEMBER_MEMORY;
 
 public class Hibernate extends CustomJorbsModCard {
     public static final String ID = JorbsMod.makeID(Hibernate.class.getSimpleName());
