+++ src/main/java/stsjorbsmod/cards/wanderer/HoldMonster.java
@@ -14,7 +14,7 @@ import stsjorbsmod.characters.Wanderer;
 import stsjorbsmod.memories.KindnessMemory;
 
 import static stsjorbsmod.JorbsMod.makeCardPath;
-import static stsjorbsmod.characters.Wanderer.Enums.REMEMBER_MEMORY;
+import static stsjorbsmod.JorbsMod.JorbsCardTags.REMEMBER_MEMORY;
 
 public class HoldMonster extends CustomJorbsModCard {
     public static final String ID = JorbsMod.makeID(HoldMonster.class.getSimpleName());
