+++ src/main/java/stsjorbsmod/cards/wanderer/MirrorImage.java
@@ -12,7 +12,7 @@ import stsjorbsmod.memories.KindnessMemory;
 import stsjorbsmod.powers.NextAttackMissesPower;
 
 import static stsjorbsmod.JorbsMod.makeCardPath;
-import static stsjorbsmod.characters.Wanderer.Enums.REMEMBER_MEMORY;
+import static stsjorbsmod.JorbsMod.JorbsCardTags.REMEMBER_MEMORY;
 
 public class MirrorImage extends CustomJorbsModCard {
     public static final String ID = JorbsMod.makeID(MirrorImage.class.getSimpleName());
