+++ src/main/java/stsjorbsmod/cards/wanderer/Rest.java
@@ -11,7 +11,7 @@ import stsjorbsmod.characters.Wanderer;
 import stsjorbsmod.memories.ChastityMemory;
 
 import static stsjorbsmod.JorbsMod.makeCardPath;
-import static stsjorbsmod.characters.Wanderer.Enums.REMEMBER_MEMORY;
+import static stsjorbsmod.JorbsMod.JorbsCardTags.REMEMBER_MEMORY;
 
 public class Rest extends CustomJorbsModCard {
     public static final String ID = JorbsMod.makeID(Rest.class.getSimpleName());
