+++ src/main/java/stsjorbsmod/cards/wanderer/TinyHut.java
@@ -10,7 +10,7 @@ import stsjorbsmod.characters.Wanderer;
 import stsjorbsmod.memories.KindnessMemory;
 
 import static stsjorbsmod.JorbsMod.makeCardPath;
-import static stsjorbsmod.characters.Wanderer.Enums.REMEMBER_MEMORY;
+import static stsjorbsmod.JorbsMod.JorbsCardTags.REMEMBER_MEMORY;
 
 public class TinyHut extends CustomJorbsModCard {
     public static final String ID = JorbsMod.makeID(TinyHut.class.getSimpleName());
