+++ src/main/java/stsjorbsmod/cards/wanderer/WanderingMind.java
@@ -9,7 +9,7 @@ import stsjorbsmod.actions.RememberRandomNewMemoryAction;
 import stsjorbsmod.characters.Wanderer;
 
 import static stsjorbsmod.JorbsMod.makeCardPath;
-import static stsjorbsmod.characters.Wanderer.Enums.REMEMBER_MEMORY;
+import static stsjorbsmod.JorbsMod.JorbsCardTags.REMEMBER_MEMORY;
 
 public class WanderingMind extends CustomJorbsModCard {
     public static final String ID = JorbsMod.makeID(WanderingMind.class.getSimpleName());
