+++ src/main/java/stsjorbsmod/cards/wanderer/WeightOfMemory.java
@@ -12,7 +12,7 @@ import stsjorbsmod.cards.CustomJorbsModCard;
 import stsjorbsmod.characters.Wanderer;
 
 import static stsjorbsmod.JorbsMod.makeCardPath;
-import static stsjorbsmod.characters.Wanderer.Enums.REMEMBER_MEMORY;
+import static stsjorbsmod.JorbsMod.JorbsCardTags.REMEMBER_MEMORY;
 
 public class WeightOfMemory extends CustomJorbsModCard {
     public static final String ID = JorbsMod.makeID(WeightOfMemory.class.getSimpleName());
